(maint) rubocop autocorrect
This addresses more than half the complaints from rubocop. The only
necessary manual change was a `unless` condition in the Rakefile.

+++ lib/pdk/module/templatedir.rb
@@ -34,7 +34,7 @@ module PDK
       # @raise [ArgumentError] (see #validate_module_template!)
       #
       # @api public
-      def initialize(path_or_url, &block)
+      def initialize(path_or_url)
         if File.directory?(path_or_url)
           @path = path_or_url
         else
@@ -50,7 +50,7 @@ module PDK
           unless clone_result[:exit_code] == 0
             PDK.logger.error clone_result[:stdout]
             PDK.logger.error clone_result[:stderr]
-            raise PDK::CLI::FatalError, _("Unable to clone git repository '%{repo}' to '%{dest}'") % {:repo => path_or_url, :dest => temp_dir}
+            raise PDK::CLI::FatalError, _("Unable to clone git repository '%{repo}' to '%{dest}'") % { repo: path_or_url, dest: temp_dir }
           end
           @path = temp_dir
           @repo = path_or_url
@@ -63,9 +63,7 @@ module PDK
       ensure
         # If we cloned a git repo to get the template, remove the clone once
         # we're done with it.
-        if @repo
-          FileUtils.remove_dir(@path)
-        end
+        FileUtils.remove_dir(@path) if @repo
       end
 
       # Retrieve identifying metadata for the template.
@@ -80,7 +78,7 @@ module PDK
         if @repo
           ref_result = PDK::CLI::Exec.git('--git-dir', File.join(@path, '.git'), 'describe', '--all', '--long')
           if ref_result[:exit_code] == 0
-            {'template-url' => @repo, 'template-ref' => ref_result[:stdout].strip}
+            { 'template-url' => @repo, 'template-ref' => ref_result[:stdout].strip }
           else
             {}
           end
@@ -100,25 +98,27 @@ module PDK
       # @return [void]
       #
       # @api public
-      def render(&block)
+      def render
         files_in_template.each do |template_file|
-          PDK.logger.debug(_("Rendering '%{template}'...") % {:template => template_file})
+          PDK.logger.debug(_("Rendering '%{template}'...") % { template: template_file })
           dest_path = template_file.sub(/\.erb\Z/, '')
 
           begin
-            dest_content = PDK::TemplateFile.new(File.join(@moduleroot_dir, template_file), {:configs => config_for(dest_path)}).render
+            dest_content = PDK::TemplateFile.new(File.join(@moduleroot_dir, template_file), configs: config_for(dest_path)).render
           rescue => e
             error_msg = _(
-              "Failed to render template '%{template}'\n" +
-              "%{exception}: %{message}"
-              ) % {:template => template_file, :exception => e.class, :message => e.message}
+              "Failed to render template '%{template}'\n" \
+              '%{exception}: %{message}'
+            ) % { template: template_file, exception: e.class, message: e.message }
             raise PDK::CLI::FatalError, error_msg
           end
 
           yield dest_path, dest_content
         end
       end
-    private
+
+      private
+
       # Validate the content of the template directory.
       #
       # @raise [ArgumentError] If the specified path is not a directory.
@@ -130,11 +130,11 @@ module PDK
       # @api private
       def validate_module_template!
         unless File.directory?(@path)
-          raise ArgumentError, _("The specified template '%{path}' is not a directory") % {:path => @path}
+          raise ArgumentError, _("The specified template '%{path}' is not a directory") % { path: @path }
         end
 
         unless File.directory?(@moduleroot_dir)
-          raise ArgumentError, _("The template at '%{path}' does not contain a moduleroot directory") % {:path => @path}
+          raise ArgumentError, _("The template at '%{path}' does not contain a moduleroot directory") % { path: @path }
         end
       end
 
@@ -145,11 +145,11 @@ module PDK
       #
       # @api private
       def files_in_template
-        @files ||= Dir.glob(File.join(@moduleroot_dir, "**", "*"), File::FNM_DOTMATCH).select { |template_path|
+        @files ||= Dir.glob(File.join(@moduleroot_dir, '**', '*'), File::FNM_DOTMATCH).select do |template_path|
           File.file?(template_path) && !File.symlink?(template_path)
-        }.map { |template_path|
+        end.map do |template_path|
           template_path.sub(/\A#{Regexp.escape(@moduleroot_dir)}#{Regexp.escape(File::SEPARATOR)}/, '')
-        }
+        end
       end
 
       # Generate a hash of data to be used when rendering the specified
@@ -172,9 +172,9 @@ module PDK
 
           if File.file?(config_path) && File.readable?(config_path)
             begin
-              @config = YAML.load(File.read(config_path))
+              @config = YAML.safe_load(File.read(config_path))
             rescue
-              PDK.logger.warn(_("'%{file}' is not a valid YAML file") % {:file => config_path})
+              PDK.logger.warn(_("'%{file}' is not a valid YAML file") % { file: config_path })
               @config = {}
             end
           else
